# backend/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/folders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "List folders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/folder.Folder"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve folders",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a folder owned by the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Create a folder",
                "parameters": [
                    {
                        "description": "Folder payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/folder.CreateFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/folder.Folder"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/folders/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Rename a folder",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/folder.UpdateFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/folder.Folder"
                        }
                    },
                    "400": {
                        "description": "Invalid folder ID or payload",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Folder not found",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes the folder and every document it contains.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Delete a folder",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/folder.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid folder ID",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete folder",
                        "schema": {
                            "$ref": "#/definitions/folder.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "folder.CreateFolderRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Contracts"
                }
            }
        },
        "folder.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to delete folder"
                }
            }
        },
        "folder.Folder": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "folder.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Folder deleted"
                }
            }
        },
        "folder.UpdateFolderRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Contracts 2026"
                }
            }
        }
    }
}